Added retry and ratelimit

+++ scrapper/src/main/java/edu/java/controller/LinksController.java
@@ -1,5 +1,6 @@
 package edu.java.controller;
 
+import edu.java.controller.buckets.Limiter;
 import edu.java.controller.dto.AddLinkRequest;
 import edu.java.controller.dto.ApiErrorResponse;
 import edu.java.controller.dto.LinkResponse;
@@ -13,9 +14,9 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
 import io.swagger.v3.oas.annotations.tags.Tag;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import java.util.List;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -29,11 +30,13 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping("/links")
 @Tag(name = "Links", description = "API for working with links")
-public class LinksController {
+public class LinksController extends BaseController {
     private final LinkService linkService;
+    private final Limiter limiter;
 
-    public LinksController(LinkService linkService) {
+    public LinksController(LinkService linkService, Limiter limiter) {
         this.linkService = linkService;
+        this.limiter = limiter;
     }
 
     @GetMapping
@@ -47,12 +50,19 @@ public class LinksController {
         }),
         @ApiResponse(responseCode = "404", description = "Chat not found", content = {
             @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
+        }),
+        @ApiResponse(responseCode = "429", description = "Too many requests", content = {
+            @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
         })
     })
-    public ResponseEntity<ListLinksResponse> getLinks(@RequestHeader("Tg-Chat-Id") long chatId) {
-        List<LinkResponse> links = linkService.listAll(chatId).stream().map(link -> new LinkResponse(
-            link.getId(), link.getUrl())).toList();
-        return ResponseEntity.ok().body(new ListLinksResponse(links, links.size()));
+    public ResponseEntity<?> getLinks(@RequestHeader("Tg-Chat-Id") long chatId, HttpServletRequest request) {
+        if (limiter.tryConsume(request.getRemoteAddr())) {
+            List<LinkResponse> links = linkService.listAll(chatId).stream().map(link -> new LinkResponse(
+                link.getId(), link.getUrl())).toList();
+            return ResponseEntity.ok().body(new ListLinksResponse(links, links.size()));
+        } else {
+            return createTooManyRequests();
+        }
     }
 
     @PostMapping
@@ -73,18 +83,26 @@ public class LinksController {
         }),
         @ApiResponse(responseCode = "409", description = "Link is already tracking", content = {
             @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
+        }),
+        @ApiResponse(responseCode = "429", description = "Too many requests", content = {
+            @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
         })
     })
     public ResponseEntity<?> addLink(
         @RequestHeader("Tg-Chat-Id") long chatId,
         @Valid @RequestBody AddLinkRequest addLinkRequest,
-        BindingResult bindingResult
+        BindingResult bindingResult,
+        HttpServletRequest request
     ) {
-        if (bindingResult.hasErrors()) {
-            return createBadRequestResponse(bindingResult);
+        if (limiter.tryConsume(request.getRemoteAddr())) {
+            if (bindingResult.hasErrors()) {
+                return createBadRequestResponse(bindingResult);
+            }
+            Link addedLink = linkService.add(addLinkRequest.link(), chatId);
+            return ResponseEntity.ok().body(new LinkResponse(addedLink.getId(), addedLink.getUrl()));
+        } else {
+            return createTooManyRequests();
         }
-        Link addedLink = linkService.add(addLinkRequest.link(), chatId);
-        return ResponseEntity.ok().body(new LinkResponse(addedLink.getId(), addedLink.getUrl()));
     }
 
     @DeleteMapping
@@ -105,25 +123,25 @@ public class LinksController {
                      content = {
                          @Content(mediaType = "application/json",
                                   schema = @Schema(implementation = ApiErrorResponse.class))
-                     })
+                     }),
+        @ApiResponse(responseCode = "429", description = "Too many requests", content = {
+            @Content(mediaType = "application/json", schema = @Schema(implementation = ApiErrorResponse.class))
+        })
     })
     public ResponseEntity<?> deleteLink(
         @RequestHeader("Tg-Chat-Id") long chatId,
         @Valid @RequestBody RemoveLinkRequest removeLinkRequest,
-        BindingResult bindingResult
+        BindingResult bindingResult,
+        HttpServletRequest request
     ) {
-        if (bindingResult.hasErrors()) {
-            return createBadRequestResponse(bindingResult);
+        if (limiter.tryConsume(request.getRemoteAddr())) {
+            if (bindingResult.hasErrors()) {
+                return createBadRequestResponse(bindingResult);
+            }
+            Link deleted = linkService.remove(removeLinkRequest.link(), chatId);
+            return ResponseEntity.ok().body(new LinkResponse(deleted.getId(), deleted.getUrl()));
+        } else {
+            return createTooManyRequests();
         }
-        Link deleted = linkService.remove(removeLinkRequest.link(), chatId);
-        return ResponseEntity.ok().body(new LinkResponse(deleted.getId(), deleted.getUrl()));
-    }
-
-    private ResponseEntity<ApiErrorResponse> createBadRequestResponse(BindingResult bindingResult) {
-        return ResponseEntity.badRequest()
-            .body(new ApiErrorResponse(
-                bindingResult.getAllErrors().getFirst().getDefaultMessage(),
-                Integer.toString(HttpStatus.BAD_REQUEST.value())
-            ));
     }
 }
